chore: refactor updateUserProfile
