+++ packages/server/graphql/mutations/uploadOrgImage.ts
@@ -5,7 +5,7 @@ import normalizeAvatarUpload from '../../fileStorage/normalizeAvatarUpload'
 import validateAvatarUpload from '../../fileStorage/validateAvatarUpload'
 import {getUserId, isUserBillingLeader} from '../../utils/authorization'
 import standardError from '../../utils/standardError'
-import GraphQLFileType from '../types/GraphQLFileType'
+import GraphQLFileType from '../public/types/GraphQLFileType'
 import UpdateOrgPayload from '../types/UpdateOrgPayload'
 import {GQLContext} from './../graphql'
 import {default as updateOrgResolver} from './helpers/updateOrg'
