+++ packages/server/graphql/public/typeDefs/updateUserProfile.graphql
@@ -1,0 +1,25 @@
+extend type Mutation {
+  updateUserProfile(
+    """
+    The input object containing the user profile fields that can be changed
+    """
+    updatedUser: UpdateUserProfileInput!
+  ): UpdateUserProfilePayload
+}
+
+input UpdateUserProfileInput {
+  """
+  The name, as confirmed by the user
+  """
+  preferredName: String
+}
+
+type UpdateUserProfilePayload {
+  error: StandardMutationError
+  user: User
+
+  """
+  The updated team member
+  """
+  teamMembers: [TeamMember!]
+}
