+++ packages/server/graphql/public/typeDefs/uploadUserImage.graphql
@@ -1,0 +1,11 @@
+extend type Mutation {
+  """
+  Upload an image for a user avatar
+  """
+  uploadUserImage(
+    """
+    the user avatar image file
+    """
+    file: File!
+  ): UpdateUserProfilePayload
+}
