+++ packages/server/graphql/public/types/GraphQLFileType.ts
@@ -1,5 +1,13 @@
 import {GraphQLScalarType} from 'graphql'
 
+export type File = {
+  contentType: string
+  buffer: {
+    type: 'Buffer'
+    data: Array<number>
+  }
+}
+
 const GraphQLFileType = new GraphQLScalarType({
   name: 'File',
   serialize: (value) => value,
