+++ packages/server/graphql/rootMutation.ts
@@ -125,10 +125,8 @@ import updateTask from './mutations/updateTask'
 import updateTaskDueDate from './mutations/updateTaskDueDate'
 import updateTeamName from './mutations/updateTeamName'
 import updateTemplateScope from './mutations/updateTemplateScope'
-import updateUserProfile from './mutations/updateUserProfile'
 import upgradeToTeamTier from './mutations/upgradeToTeamTier'
 import uploadOrgImage from './mutations/uploadOrgImage'
-import uploadUserImage from './mutations/uploadUserImage'
 import voteForPokerStory from './mutations/voteForPokerStory'
 import voteForReflectionGroup from './mutations/voteForReflectionGroup'
 
@@ -243,10 +241,8 @@ export default new GraphQLObjectType<any, GQLContext>({
       updateTaskDueDate,
       updateTeamName,
       updateTemplateScope,
-      updateUserProfile,
       upgradeToTeamTier,
       uploadOrgImage,
-      uploadUserImage,
       voteForReflectionGroup,
       voteForPokerStory,
       pokerRevealVotes,
